Refresh admin overview on event deletion and add invite code column
Adds a column to the management overview event table with the invite code.
Adds an option to the context menu to copy an event's invite code to the clipboard.
Reload the list of events after the deletion of an event (after the server tells us it was deleted).
Adds a confirmation dialog before deleting an event.